Framed protocol
